Initial commit of Mutiny Editor.
Note: The Editor is quite low priority for this project because as with the commercial Unity editor, it is often faster and more robust for a technical developer to do much of the work via code using the API directly.

The Mutiny Editor is mainly going to be for laying out scenes rather than assigning data to public variables in scripts. However, I will probably end up experimenting with expanding the code generation for scenes to also manipulate behaviours at some point. Could be fun ;)